CheckEasyBackend/requirements.txt
# File: requirements.txt
# 以下依赖锁定了项目所需的关键依赖版本，确保在 Docker 环境中安装的版本稳定兼容。
# 其中部分包版本来自于当前环境（通过 pip freeze 获取），但对 Pillow、pytesseract 和 spacy 进行了覆盖，
# 并移除了 uvicorn（不需要）。

(venv) cjh@Mac CheckEasy % pip freeze
aiofiles==24.1.0
alembic==1.15.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==3.2.2
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
emoji==2.14.1
fastapi==0.115.11
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
imageio==2.37.0
iniconfig==2.0.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
kombu==5.4.2
lazy_loader==0.4
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
passlib==1.7.4
PassportEye==2.2.2
pdfminer==20191125
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
protobuf==6.30.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.1
pytesseract==0.3.13
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
stanza==1.10.1
starlette==0.46.0
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2025.2.18
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13 